+++ projects/hipblaslt/clients/tests/src/matmul_gtest.cpp
@@ -157,12 +157,6 @@ namespace
                 if(arg.scaleAlpha_vector)
                     name << "_SAV";
 
-                if(arg.amaxScaleA)
-                    name << "_ASA";
-
-                if(arg.amaxScaleB)
-                    name << "_ASB";
-
                 if(arg.amaxD)
                     name << "_AMaxD";